requirements.txt
Package                       Version
----------------------------- -----------
aliyun-python-sdk-core        2.13.26
aliyun-python-sdk-slb         3.2.18
amqp                          2.6.1
ansible                       2.9.13
atomicwrites                  1.4.0
attrs                         20.2.0
backports.functools-lru-cache 1.6.1
billiard                      3.6.3.0
celery                        4.3.1
celery-once                   3.0.1
celery-redbeat                0.13.0
cffi                          1.14.2
click                         7.1.2
configparser                  4.0.2
contextlib2                   0.6.0.post1
cryptography                  3.1
enum34                        1.1.10
Flask                         1.1.1
Flask-CLI                     0.4.0
Flask-Cors                    3.0.9
Flask-MySQLdb                 0.2.0
Flask-SQLAlchemy              2.4.4
funcsigs                      1.0.2
futures                       3.3.0
importlib-metadata            1.7.0
ipaddress                     1.0.23
itsdangerous                  1.1.0
Jinja2                        2.11.2
jmespath                      0.10.0
jsonpatch                     1.26
jsonpointer                   2.0
kombu                         4.6.11
MarkupSafe                    1.1.1
monotonic                     1.5
more-itertools                5.0.0
mysqlclient                   1.4.6
packaging                     20.4
pathlib2                      2.3.5
pip                           20.2.2
pluggy                        0.13.1
py                            1.9.0
pycparser                     2.20
pycryptodome                  3.9.8
pyparsing                     2.4.7
pytest                        4.6.11
python-dateutil               2.8.1
python-dotenv                 0.14.0
pytz                          2020.1
PyYAML                        5.3.1
redis                         3.5.3
scandir                       1.10.0
setuptools                    44.1.1
six                           1.15.0
SQLAlchemy                    1.3.19
tenacity                      6.2.0
typing                        3.7.4.3
vine                          1.3.0
wcwidth                       0.2.5
Werkzeug                      1.0.1
wheel                         0.35.1
zipp                          1.2.0
